third_party/blink/web_tests/svg/dom/SVGStyleElement/disable-svg-style-element-expected.txt

Test the disabled property on an svg style element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS styleElement.disabled is false
PASS window.getComputedStyle(rect).fill is "rgb(0, 0, 255)"
PASS styleElement.disabled is true
PASS window.getComputedStyle(rect).fill is "rgb(255, 0, 0)"
PASS newStyleElement.disabled is false
PASS newStyleElement.disabled is false
PASS otherStyleElement.disabled is false
PASS otherStyleElement.disabled is false
PASS successfullyParsed is true

TEST COMPLETE
